Move Card.js to components folder, use in Login
* Move the Card.js (and .css) to components folder
* Replace html in Login.js to directly use Card rather than just the
class names
