+++ expo/src/components/Card.css
@@ -1,5 +1,5 @@
 
-@import url('./customize/customize.css');
+@import url('../customize/customize.css');
 /* TODO decide on app-wide color scheme */
 
 .card {
